% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{Node}
\alias{Node}
\title{Node class for cyTree}
\description{
Represents a tree node with name, ID, parent, children, and marker annotations.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Node-new}{\code{Node$new()}}
\item \href{#method-Node-add_child}{\code{Node$add_child()}}
\item \href{#method-Node-get_path}{\code{Node$get_path()}}
\item \href{#method-Node-to_list}{\code{Node$to_list()}}
\item \href{#method-Node-print_node}{\code{Node$print_node()}}
\item \href{#method-Node-find_by_id}{\code{Node$find_by_id()}}
\item \href{#method-Node-find_by_name}{\code{Node$find_by_name()}}
\item \href{#method-Node-to_dataframes}{\code{Node$to_dataframes()}}
\item \href{#method-Node-clone}{\code{Node$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-new"></a>}}
\if{latex}{\out{\hypertarget{method-Node-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$new(
  id,
  name,
  parent = NULL,
  positive_markers = list(),
  negative_markers = list()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-add_child"></a>}}
\if{latex}{\out{\hypertarget{method-Node-add_child}{}}}
\subsection{Method \code{add_child()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$add_child(child_node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-get_path"></a>}}
\if{latex}{\out{\hypertarget{method-Node-get_path}{}}}
\subsection{Method \code{get_path()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$get_path()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-Node-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$to_list()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-print_node"></a>}}
\if{latex}{\out{\hypertarget{method-Node-print_node}{}}}
\subsection{Method \code{print_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$print_node(indent = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-find_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-Node-find_by_id}{}}}
\subsection{Method \code{find_by_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$find_by_id(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-find_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-Node-find_by_name}{}}}
\subsection{Method \code{find_by_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$find_by_name(name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-to_dataframes"></a>}}
\if{latex}{\out{\hypertarget{method-Node-to_dataframes}{}}}
\subsection{Method \code{to_dataframes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$to_dataframes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Node-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Node-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Node$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
